services/routers/events.py
from fastapi import APIRouter
from pydantic import BaseModel

router = APIRouter()

class EventRegistration(BaseModel):
    event_id: str
    user_name: str
    user_email: str

@router.get("/search")
async def search_events(type: str = "", location: str = ""):
    """Search for events (placeholder)"""
    
    # TODO: Implement with your data source
    return {
        "events": [
            {
                "id": "evt_001",
                "name": "Tech Summit 2025",
                "type": "tech",
                "location": "San Francisco",
                "date": "2025-11-15"
            },
            {
                "id": "evt_002",
                "name": "Healthcare Innovation Conference",
                "type": "healthcare",
                "location": "Boston",
                "date": "2025-12-01"
            }
        ]
    }

@router.post("/register")
async def register_event(registration: EventRegistration):
    """Register for an event (placeholder)"""
    
    # TODO: Implement actual registration logic
    return {
        "success": True,
        "message": f"✅ {registration.user_name} registered for event {registration.event_id}"
    }