services/requirements.txt
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-api-python-client
google-cloud-storage
requests
python-dotenv